Lots of little bug fixes, new user modification, new login design, lots of validation fixes

+++ BusinessBackend/src/test/java/com/main/WarehousingTests/RouteTest/RouteControllerTest.java
@@ -67,7 +67,7 @@ public class RouteControllerTest {
     public Warehouse createWarehouse(Long id){
         Warehouse warehouse = new Warehouse();
         warehouse.setId(id);
-        warehouse.setZipCode(BigDecimal.valueOf(1111));
+        warehouse.setZipCode(1111);
         warehouse.setCity("Budapest");
         warehouse.setStreet("Deák út");
         warehouse.setStreetNumber("12");
